Update international example
Setting the return value of keyboard.available() equal to "code" seemed inconsistent with examples in the PS2KeyAdvanced, and was rather confusing since (if I understand correctly, available() returns the # of key presses in the buffer, and not a code for a specific key).  

Recommend using the if( keyboard.available( ) ) construct instead.  Just my 2 cents.  Great libraries!  Thank you!